Refactored and fixed model conversion for fpa2bv conversion of unspecified values via theory_fpa.
